chore: 점수 로직 수정

+++ src/main/kotlin/com/github/kusitms_bugi/domain/session/domain/Session.kt
@@ -26,6 +26,6 @@ interface SessionStatusHistoryField<SESSION> : BaseField {
 
 interface SessionMetricField<SESSION> : BaseField {
     var session: SESSION
-    var score: Double
+    var score: Int
     var timestamp: LocalDateTime
 }
